requirements.txt
pygame>=2.5
python-dotenv>=1.0.0